moved user controller
